backend/pyproject.toml
[project]
name = "rehearsed-multi-student"
version = "0.1.0"
description = ""
authors = [
    {name = "meg dbz", email = "meganzinka@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0,<0.123.0",
    "uvicorn>=0.38.0,<0.39.0",
    "pyyaml>=6.0.3,<7.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "google-genai>=1.52.0,<2.0.0",
    "google-cloud-texttospeech>=2.14.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0"
]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
httpx = ">=0.27.0"
pytest-asyncio = ">=0.24.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
